pyproject.toml
[tool.poetry]
name = "ascalia-demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
apscheduler = "^3.10.4"
aioftp = "^0.22.3"
sqlalchemy = ">=1.4.42,<1.5"
databases = "^0.8.0"
aiosqlite = "^0.19.0"
pyyaml = "^6.0.1"
requests = "^2.31.0"
black = "^24.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
